move HostsFile type away from service definition
